Count cast to int

+++ database/zmon/20_api/05_stored_procedures/55_grafana_sprocs.sql
@@ -48,7 +48,7 @@ $$ LANGUAGE SQL VOLATILE SECURITY DEFINER;
 
 CREATE OR REPLACE FUNCTION get_tags_with_count(OUT tag TEXT, OUT "count" INT) RETURNS SETOF RECORD AS
 $$
-select tag_name, count(1) AS "count" from
+select tag_name, count(1)::INT AS "count" from
           (select json_array_elements_text((gd_dashboard->'tags')::json) AS tag_name
              from zzm_data.grafana_dashboard) t
             group by 1;
